cleaning up config and keymaps
