--- keyboards/ymdk_bface/keymaps/sethlibra/rules.mk
@@ -1,12 +1,0 @@
-
-# build options
-BOOTMAGIC_ENABLE = no
-MOUSEKEY_ENABLE = yes
-EXTRAKEY_ENABLE = yes
-CONSOLE_ENABLE = no
-COMMAND_ENABLE = yes
-BACKLIGHT_ENABLE = yes
-BACKLIGHT_CUSTOM_DRIVER = yes
-BACKLIGHT_BREATHING = no
-RGBLIGHT_ENABLE = no
-RGBLIGHT_CUSTOM_DRIVER = no
