--- keyboards/ymdk_bface/keymaps/ymdk/config.h
@@ -1,16 +1,0 @@
-#include "../../config.h"
-
-#define GRAVE_ESC_ALT_OVERRIDE
-
-#if 0
-
-/* matrix size */
-#define MATRIX_ROWS 5
-#define MATRIX_COLS 15
-
-//                        0   1   2   3   4   5   6   7   8   9   A   B   C   D   E
-#define MATRIX_ROW_PINS { B7, B6, B5, B4, B3}
-#define MATRIX_COL_PINS { A0, A1, A2, A3, A4, A5, A6, A7, C7, C6, C5, C4, C3, C2, D7}
-#define UNUSED_PINS
-
-#endif
